Use floats instead of ints for detz_offset in CSPADs

+++ xfel/command_line/xtc_process.py
@@ -119,7 +119,7 @@ xtc_phil_str = '''
     }
     cbf {
       detz_offset = None
-        .type = int
+        .type = float
         .help = Distance from back of detector rail to sample interaction region (CXI) \
                 or actual detector distance (XPP)
       override_energy = None
